sensor-simulation/requirements.txt
azure-iot-device>=2.12.0
asyncio>=3.4.3
